Add a document that defines what we're looking for in a developer
We interview and evaluate candidates regularly, we want everyone
to take part in this process. To facilitate this, we should ensure
that we have a shared collective understanding of everything that's
involved in evaluating someone. That starts with what we're looking
for.

This document contains the list of characteristics that we think are
most important in a developer that wishes to join pebble. This should
help everyone feel more comfortable when evaluating someone.